requirements.txt
Flask==2.0.1
redis==4.5.1
requests==2.28.2
python-dotenv==1.0.0
Flask-Limiter==3.3.0